+++ backend/api/__init__.py
@@ -69,11 +69,12 @@ def create_app(test_config=None):
     Migrate(app, db)
 
     # import and register blueprints
-    from api.views import main, auth
+    from api.views import main, auth, appointment
 
     # why blueprints http://flask.pocoo.org/docs/1.0/blueprints/
     app.register_blueprint(main.main)
     app.register_blueprint(auth.auth)
+    app.register_blueprint(appointment.appointment)
 
     # register error Handler
     app.register_error_handler(Exception, all_exception_handler)
